feat: Allow the worker code to process an arbitrary number of tasks at a time (#187)
* feat: Allow the worker code to process an arbitrary number of tasks at a time

This solves the case where we are effectively just looking to shard the tasks to different workers.
People have the flexiibility in their consumer code to configure how things work

* docs: Various changes

Update Celery and Modal examples.
Update references to task_batch_size in the docs.
Additionally provide an explicit section with discussion of queue usage